refactor(sqlite): isolate task ledger schema
